Refactor s3stalker to comply with review
